move config file to appropriate path (#1408)
