+++ lib/cryptoexchange.rb
@@ -14,6 +14,7 @@ require "cryptoexchange/services/market_stream"
 require "cryptoexchange/services/trade_stream"
 require "cryptoexchange/services/order_book_stream"
 require "cryptoexchange/services/authentication"
+require "cryptoexchange/config"
 
 path_files = Dir[File.join(File.dirname(__dir__), 'lib', 'cryptoexchange', '**', '*.rb')]
 
@@ -21,9 +22,8 @@ path_files.sort_by!{ |file_name| file_name.downcase }.each do |path|
   require_relative path
 end
 
+Cryptoexchange.reset_config
+
 require "http"
 require "lru_redux"
 require "websocket-eventmachine-client"
-require "config"
-
-Cryptoexchange.reset_config
