+++ lib/cryptoexchange/config.rb
